extract useSelectedTextStyles to seperate hook
